+++ cyclo_control/CHANGELOG.rst
@@ -1,6 +1,14 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package cyclo_motion_controller
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Changelog for package cyclo_control
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+0.1.2 (2026-04-06)
+------------------
+* Removed gripper joint from the trajectory messages
+* Renamed meta package to cyclo_control
+* Changed reactivate service to topic
+* Contributors: Yeonguk Kim
+
 0.1.1 (2026-03-30)
 ------------------
 * Added acknowledgements for the dependencies used in the project