+++ cyclo_motion_controller_ros/CHANGELOG.rst
@@ -1,6 +1,13 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 Changelog for package cyclo_motion_controller_ros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+0.1.2 (2026-04-06)
+------------------
+* Removed gripper joint from the trajectory messages
+* Changed reactivate service to topic
+* Contributors: Yeonguk Kim
+
 0.1.1 (2026-03-30)
 ------------------
 * None